feat: refactor HomeScreen into main.dart for proper visibility on pub.dev example tab
